Lots of methods: set_instances and select_demos.
set_instances lives in the resumes controller and is responsible for
setting up the @resume, @experiences, and @links instances.

select_demos lives in the experience.rb model file and is responsible
for grabbing the right demos depending upon the resume_name parameter
(which is consequently eventually going to be the route).